DAT-1136-bugfix export from links
