+++ cmdb/interface/dispatcher_middleware.py
@@ -1,0 +1,54 @@
+# DATAGERRY - OpenSource Enterprise CMDB
+# Copyright (C) 2024 becon GmbH
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see <https://www.gnu.org/licenses/>.
+"""
+Server module for web-based services
+"""
+import logging
+
+from cmdb import __MODE__
+# -------------------------------------------------------------------------------------------------------------------- #
+
+LOGGER = logging.getLogger(__name__)
+
+# -------------------------------------------------------------------------------------------------------------------- #
+#                                             DispatcherMiddleware - CLASS                                             #
+# -------------------------------------------------------------------------------------------------------------------- #
+class DispatcherMiddleware:
+    """TODO: document"""
+
+    def __init__(self, app, mounts=None):
+        self.app = app
+        self.mounts = mounts or {}
+
+
+    def __call__(self, environ, start_response):
+        script = environ.get('PATH_INFO', '')
+        path_info = ''
+
+        while '/' in script:
+            if script in self.mounts:
+                app = self.mounts[script]
+                break
+            script, last_item = script.rsplit('/', 1)
+            path_info = f'/{last_item}{path_info}'
+        else:
+            app = self.mounts.get(script, self.app)
+
+        original_script_name = environ.get('SCRIPT_NAME', '')
+        environ['SCRIPT_NAME'] = original_script_name + script
+        environ['PATH_INFO'] = path_info
+
+        return app(environ, start_response)
