+++ cmdb/interface/rest_api/routes/exporter_routes/exporter_object_routes.py
@@ -63,6 +63,8 @@ def export_objects(params: CollectionParameters, request_user: UserModel):
         exporter = BaseExportWriter(exporter_class, _config)
 
         exporter.from_database(current_app.database_manager, request_user, AccessControlPermission.READ)
+
+        return exporter.export()
     except TypeNotFoundError as err:
         LOGGER.debug("[export_objects] TypeNotFoundError: %s", err)
         #TODO: ERROR-FIX
@@ -74,5 +76,3 @@ def export_objects(params: CollectionParameters, request_user: UserModel):
     except Exception as err:
         LOGGER.debug("[export_objects] Exception: %s, Type: %s", err, type(err))
         return abort(404, jsonify(message='Not Found', error='Export objects Exception'))
-
-    return exporter.export()
