+++ cmdb/interface/rest_api/routes/settings_routes/system_routes.py
@@ -21,7 +21,7 @@ import logging
 from cmdb.manager.manager_provider_model.manager_provider import ManagerProvider
 from cmdb.manager.manager_provider_model.manager_type_enum import ManagerType
 from cmdb.manager.settings_reader_manager import SettingsReaderManager
-from cmdb.utils.system_config import SystemConfigReader
+from cmdb.utils.system_config_reader import SystemConfigReader
 
 from cmdb import __title__, __version__, __runtime__
 from cmdb.interface.rest_api.routes.framework_routes.setting_routes import settings_blueprint
