+++ cmdb/startup_routines/setup_routine.py
@@ -24,7 +24,7 @@ from cmdb.manager.users_manager import UsersManager
 
 from cmdb.startup_routines.setup_status_enum import SetupStatus
 from cmdb.models.user_model.user import UserModel
-from cmdb.utils.system_config import SystemConfigReader
+from cmdb.utils.system_config_reader import SystemConfigReader
 from cmdb.manager.settings_writer_manager import SettingsWriterManager
 from cmdb.updater.updater_module import UpdaterModule
 from cmdb.updater.updater_settings import UpdateSettings
